generalizing quality check (works for UCONN)
